// src/app/modelo/enum/dominios-enum.ts
 
export enum dominiosEnum {
    Entidad = 1,
    TipoParametros = 41,
    UnidadesMedida = 62,
    Metodos = 35,
    Variables  = 107,
    Periodos = 103,
    ZonaOperativaEAAB = 25,
    TipoPozo = 121,
    CategoriaPozo = 122,
    CondicionPozo = 123,
    CategoriasEstacion = 21,
    TecnologiasEstacion = 22,
    TipoEstacion = 23,
    EstadoEstacion = 24, 
    AreaHidrografica = 28,
    ZonaHidrografica = 29,
    Cuenca = 30,
    MicroCuenca = 31,
    TipoArchivo = 34,
    TipoFraccionamientoArchvios = 101,
    UbicacionDatos = 102,
    FrecuenciasTemporales = 103,
    CodificacíonDatos = 104,
    TipoCoordenadas = 124,
    TipoDato = 105,
    TipoContenido = 106,
    TipoCorriente = 141,
    FlagObservacion = 163,
    TipoElemento=183,
    TipoAforos=201,
    tipoEspejo=202,
    tipoMolinete=221,
    FrecuenciasCalibracion =241,
    metodoMedicion = 283,
    PeriodoReporte =282,
    TipoReporte =281,
    TipoValidacion =162,
    TipoRelacion =321,

    
}